Split up _api.py to deparate modules.
