+++ webresource/resources.py
@@ -1,97 +1,11 @@
-from collections import Counter
 import base64
-import copy
 import hashlib
-import logging
 import os
-import sys
 import uuid
 
-
-try:
-    FileNotFoundError
-except NameError:  # pragma: nocover
-    FileNotFoundError = EnvironmentError
-
-
-logger = logging.getLogger(__name__)
-is_py3 = sys.version_info[0] >= 3
-namespace_uuid = uuid.UUID('f3341b2e-f97e-40d2-ad2f-10a08a778877')
-
-
-class ResourceConfig(object):
-    """Config singleton for web resources."""
-
-    def __init__(self):
-        self.development = False
-
-
-config = ResourceConfig()
-
-
-class ResourceError(ValueError):
-    """Resource related exception."""
-
-
-class ResourceMixin(object):
-    """Mixin for ``Resource`` and ``ResourceGroup``."""
-
-    def __init__(
-        self, name='', directory=None, path=None, include=True, group=None
-    ):
-        self.name = name
-        self.directory = directory
-        self.path = path
-        self.include = include
-        self.parent = None
-        if group:
-            group.add(self)
-
-    @property
-    def path(self):
-        if self._path is not None:
-            return self._path
-        if self.parent is not None:
-            return self.parent.path
-
-    @path.setter
-    def path(self, path):
-        self._path = path
-
-    @property
-    def directory(self):
-        if self._directory is not None:
-            return self._directory
-        if self.parent is not None:
-            return self.parent.directory
-
-    @directory.setter
-    def directory(self, directory):
-        if directory is None:
-            self._directory = None
-            return
-        self._directory = os.path.abspath(directory)
-
-    @property
-    def include(self):
-        if callable(self._include):
-            return self._include()
-        return self._include
-
-    @include.setter
-    def include(self, include):
-        self._include = include
-
-    def remove(self):
-        """Remove resource or resource group from parent group."""
-        if not self.parent:
-            raise ResourceError('Object is no member of a resource group')
-        self.parent.members.remove(self)
-        self.parent = None
-
-    def copy(self):
-        """Return a deep copy of this object."""
-        return copy.deepcopy(self)
+from webresource.base import ResourceMixin
+from webresource.config import config, is_py3, namespace_uuid
+from webresource.exceptions import ResourceError
 
 
 class Resource(ResourceMixin):
@@ -500,236 +414,3 @@ class StyleResource(LinkMixin):
             type_='text/css', hreflang=hreflang, media=media, rel=rel,
             sizes=None, title=title, **kwargs
         )
-
-
-class ResourceGroup(ResourceMixin):
-    """A resource group."""
-
-    def __init__(
-        self, name='', directory=None, path=None, include=True, group=None
-    ):
-        """Create resource group.
-
-        :param name: The resource group name.
-        :param directory: Directory containing the resource files.
-        :param path: Optional URL path for HTML tag link creation. Takes
-            precedence over group members paths.
-        :param include: Flag or callback function returning a flag whether to
-            include the resource group.
-        :param group: Optional resource group instance.
-        """
-        super(ResourceGroup, self).__init__(
-            name=name, directory=directory, path=path,
-            include=include, group=group
-        )
-        self._members = []
-
-    @property
-    def members(self):
-        """List of group members.
-
-        Group members are either instances of ``Resource`` or ``ResourceGroup``.
-        """
-        return self._members
-
-    @property
-    def scripts(self):
-        """List of all contained ``ScriptResource`` instances.
-
-        Resources from subsequent resource groups are included.
-        """
-        return self._filtered_resources(ScriptResource)
-
-    @property
-    def styles(self):
-        """List of all contained ``StyleResource`` instances.
-
-        Resources from subsequent resource groups are included.
-        """
-        return self._filtered_resources(StyleResource)
-
-    @property
-    def links(self):
-        """List of all contained ``LinkResource`` instances.
-
-        Resources from subsequent resource groups are included.
-        """
-        return self._filtered_resources(LinkResource)
-
-    def add(self, member):
-        """Add member to resource group.
-
-        :param member: Either ``ResourceGroup`` or ``Resource`` instance.
-        :raise ResourceError: Invalid member given.
-        """
-        if not isinstance(member, (ResourceGroup, Resource)):
-            raise ResourceError(
-                'Resource group can only contain instances '
-                'of ``ResourceGroup`` or ``Resource``'
-            )
-        member.parent = self
-        self._members.append(member)
-
-    def _filtered_resources(self, type_, members=None):
-        if members is None:
-            members = self.members
-        resources = []
-        for member in members:
-            if isinstance(member, ResourceGroup):
-                resources += self._filtered_resources(
-                    type_,
-                    members=member.members
-                )
-            elif isinstance(member, type_):
-                resources.append(member)
-        return resources
-
-    def __repr__(self):
-        return '{} name="{}"'.format(
-            self.__class__.__name__,
-            self.name
-        )
-
-
-class ResourceConflictError(ResourceError):
-    """Multiple resources declared with the same name."""
-
-    def __init__(self, counter):
-        conflicting = list()
-        for name, count in counter.items():
-            if count > 1:
-                conflicting.append(name)
-        msg = 'Conflicting resource names: {}'.format(sorted(conflicting))
-        super(ResourceConflictError, self).__init__(msg)
-
-
-class ResourceCircularDependencyError(ResourceError):
-    """Resources define circular dependencies."""
-
-    def __init__(self, resources):
-        msg = 'Resources define circular dependencies: {}'.format(resources)
-        super(ResourceCircularDependencyError, self).__init__(msg)
-
-
-class ResourceMissingDependencyError(ResourceError):
-    """Resource depends on a missing resource."""
-
-    def __init__(self, resource):
-        msg = 'Resource defines missing dependency: {}'.format(resource)
-        super(ResourceMissingDependencyError, self).__init__(msg)
-
-
-class ResourceResolver(object):
-    """Resource resolver."""
-
-    def __init__(self, members):
-        """Create resource resolver.
-
-        :param members: Either single or list of ``Resource`` or
-            ``ResourceGroup`` instances.
-        :raise ResourceError: Members contain invalid member.
-        """
-        if not isinstance(members, (list, tuple)):
-            members = [members]
-        for member in members:
-            if not isinstance(member, (Resource, ResourceGroup)):
-                raise ResourceError(
-                    'members can only contain instances '
-                    'of ``ResourceGroup`` or ``Resource``'
-                )
-        self.members = members
-
-    def _flat_resources(self, members=None):
-        if members is None:
-            members = self.members
-        resources = []
-        for member in members:
-            if not member.include:
-                continue
-            if isinstance(member, ResourceGroup):
-                resources += self._flat_resources(members=member.members)
-            else:
-                resources.append(member)
-        return resources
-
-    def resolve(self):
-        """Return all resources from members as flat list ordered by
-        dependencies.
-
-        :raise ResourceConflictError: Resource list contains conflicting names
-        :raise ResourceMissingDependencyError: Dependency resource not included
-        :raise ResourceCircularDependencyError: Circular dependency defined.
-        """
-        resources = self._flat_resources()
-        names = [res.name for res in resources]
-        counter = Counter(names)
-        if len(resources) != len(counter):
-            raise ResourceConflictError(counter)
-        ret = []
-        handled = {}
-        for resource in resources[:]:
-            if not resource.depends:
-                ret.append(resource)
-                handled[resource.name] = resource
-                resources.remove(resource)
-            else:
-                for dependency_name in resource.depends:
-                    if dependency_name not in names:
-                        raise ResourceMissingDependencyError(resource)
-        count = len(resources)
-        while count > 0:
-            count -= 1
-            for resource in resources[:]:
-                hook_idx = 0
-                not_yet = False
-                for dependency_name in resource.depends:
-                    if dependency_name in handled:
-                        dependency = handled[dependency_name]
-                        dep_idx = ret.index(dependency)
-                        hook_idx = dep_idx if dep_idx > hook_idx else hook_idx
-                    else:
-                        not_yet = True
-                        break
-                if not_yet:
-                    continue
-                ret.insert(hook_idx + 1, resource)
-                handled[resource.name] = resource
-                resources.remove(resource)
-                break
-        if resources:
-            raise ResourceCircularDependencyError(resources)
-        return ret
-
-
-class ResourceRenderer(object):
-    """Resource renderer."""
-
-    def __init__(self, resolver, base_url='https://tld.org'):
-        """Create resource renderer.
-
-        :param resolver: ``ResourceResolver`` instance.
-        :param base_url: Base URL to render resource HTML tags.
-        """
-        self.resolver = resolver
-        self.base_url = base_url
-
-    def render(self):
-        """Render resources."""
-        return u'\n'.join([
-            res.render(self.base_url) for res in self.resolver.resolve()
-        ])
-
-
-class GracefulResourceRenderer(ResourceRenderer):
-    """Resource renderer, which does not fail but logs an exception."""
-
-    def render(self):
-        lines = []
-        for resource in self.resolver.resolve():
-            try:
-                lines.append(resource.render(self.base_url))
-            except (ResourceError, FileNotFoundError):
-                msg = u'Failure to render resource "{}"'.format(resource.name)
-                lines.append(u'<!-- {} - details in logs -->'.format(msg))
-                logger.exception(msg)
-        return u'\n'.join(lines)
